requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.2
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
